Django_prod/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django~=4.0.2
gunicorn==20.1.0
pycodestyle==2.7.0
psycopg2==2.9.3
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
numpy~=1.20.3
pandas~=1.3.4
openpyxl~=3.0.9
djangorestframework~=3.13.1
datetime~=4.4